Fix 32-bit errors
Component type mapping changed from ulong to size_t for 32-bit
compatibility with arrays.

+++ source/star/entity/entity.d
@@ -917,7 +917,7 @@ private:
     }
 
     // Return a unique integer for every component type.
-    ulong type(C)() inout pure nothrow @safe
+    size_t type(C)() inout pure nothrow @safe
     in
     {
         assert(hasType!C());
@@ -1024,7 +1024,7 @@ private:
     Object[][] _components;
 
     // A map associating each component class with a unique unsigned integer.
-    ulong[ClassInfo] _componentTypes;
+    size_t[ClassInfo] _componentTypes;
 
     // Bitmasks of each entity's components, ordered by entity and then by component bit.
     bool[][] _componentMasks;
